requirements.txt
Django==3.1.6
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt
djangorestframework-simplejwt
psycopg2-binary==2.8.6
requests==2.24.0
pandas==1.2
python-binance==0.7.9
numpy==1.20.1
